Remove dependency on ruby-fogbugz gem
This MR removes our dependency on the `ruby-fogbugz` gem. We were
already using our own HTTP adapter and our own XML adapter, so the main
substance of this change is actually just in extracting the `Interface`
class into our own codebase and copying (with updates for style and
coverage) the relevant spec.