Link jobs to environment on creation
When a CI job references an environment (it uses the `environment`
keyword in the CI template) a new Environments::Job record is created
to persist a link to the environment.

Records are created for both Build and Bridge job types, for new jobs
and retries of existing jobs.

The new behaviour is controlled by the `persisted_job_environment_relationship`
feature flag.