Create dedicated table for storing CI environment options
Currently environment options for a CI job are stored within the job
metadata. This is not suitable because some environment options must
be kept permanently as there are features that rely on them, and
metadata is intended for short term storage.

Instead we will store these environment options in a new table, which
removes the dependency on metadata records and allows them to be deleted
as required.

Changelog: other