Allow `self_rotate` scope in API endpoint to create PAT for current user
This change set allows to use the `self_rotate` scope in addition to the
already allowed `k8s_proxy` scope for the REST API endpoint to create a
Personal Access Token for the currently authenticated user.

This change is implemented to allow clients (namely the GitLab CLI in
this case) to more efficiently manage the tokens used to authenticate
with the KAS Kubernetes API proxy.

Closes https://gitlab.com/gitlab-org/gitlab/-/issues/555546

Changelog: added