Add webhook trigger for system-initiated approval resets
This introduces a new webhook trigger for when merge requests are reset
by the system. Typically this occurs when merge requests are configured
to reset approvals on new pushes.

The new webhook payload includes a `system` field which will be `true`.
It also includes a `system_action` field which can provide more details
about the system action that triggered the reset, such as
'approvals_reset_on_push'.

Changelog: added
EE: true