Skip unnecessary EXPIRE calls when rate limiting
We store the rate limiting counts in a Redis key per interval. This key
needs to be expired after the interval ends so that we do not store
stale data.

Right now, we set the expiry on every increment. But we can optimize
this so we only set the expiry once on the first increment.