Add the ci_job_inputs table
This table will support the new job inputs feature. For each input
defined in a CI job's configuration, it stores the input name, encrypted
input value, and the type of the input. It also stores whether the input
is sensitive or not. Sensitive inputs will not be revealed in job logs.

Changelog: added