Only show draft status tag for draft MRs in list when FF is enabled
Draft status tags are behind a FF right now, but the boolean to show the
status does not take the flag into account. Users who do not have the
flag enabled see an "Open" status badge next to draft MRs in a project
MR list view (and _only_ next to drafts, not next to all open MRs). This
MR adds a check for the feature flag so that the status badge will not
show in this case.

Changelog: fixed