Update bulk_create method to also create metadata records
In order for us to deduplicate commit metadata from the
`merge_request_diff_commits` table to the new
`merge_request_commits_metadata` table, we need bulk creation
logic to also create data on the new table. This applies
to new MRDC records. A separate backfill migration will be done
along when we start to partition MRDC table as well.

This is a temporary situation. Later on as we progress in
implementing this deduplication approach, we will be removing
the logic to write commit metadata on MRDC table.

This is behind `merge_request_diff_commits_dedup` feature flag.