Move fetching merge request counts to an async call
Before this change we would fetch the users merge request count on page
load. However this query could be slow and cause 500 errors.
With this change we will only use the count on page load if it has been previously
cached, if it isn't cached we will fetch from the API which will then cache
the value to be used on next visit.

This helps improve performance for everyone in that if the cache doesn't
exist because of an action to a merge request that causes the count cache
to be invalidated the next page visit will be faster as it won't need to do this
query on page load.

https://gitlab.com/gitlab-org/gitlab/-/issues/555533