Initial extraction of PushRuleFinder
As we begin to work with push rules sourced from OrgPushRules and
GroupPushRules, we need more flexible, single-sources-of-truth that
allow us to isolate our changes and additions. Extracting
`PushRule.global` into a new finder serves this purpose, and should
serve as a behavior attractor for future development along this path.

EE: true