Reject authorize requests with checksum deploy header
Before perfoming an upload, Conan sends empty PUT requests to verify
if the file was already uploaded. These requests receive a
`411 Length required` error from workhorse, causing the upload flow
to be interrupted (non consolidated config for object storage, minio).

Since we'll handle the checksum deploy header in the `authorize`
requests then we don't really need to verify it additionally for the API
upload endpoints: it won't be present in the standard flow.