Handle MR pipelines against source branch in protected pipelines
Prior to this change, in the logic for determining if a pipeline is
protected for merge requests, the code did not account for pipelines
that run against the source branch of a merge request. It only handled
the case where the pipeline was run against the merge request reference.
The current implementation however would not cause any bug as this cause
is handled in validation when a pipeline is created.

This change fixes the issue by updating the logic to handle
the missing case where the pipeline is run against the source branch.

[Issue](https://gitlab.com/gitlab-org/gitlab/-/issues/556906)

Changelog: changed