Enhance collation checker
This MR enhances collation checker to spot check curated list of indexes
for index corruption and duplicates.
It adds migration for amcheck extension creation.
It skips the bt_index_check, if extension not enabled.

It also adds CollationCheckerWorker class. This worker gets called when
Admin runs Collation Check and saves the result in Redis cache to be
fetched later.

Changelog: changed