Delete commits metadata in Projects::DestroyService
We create records in `merge_request_commits_metadata` table which
is partitioned by `project_id`. A record in that table can
be associated to multiple `merge_request_diff_commits` record
since we re-use a row. So we don't delete it when a
`merge_request_diff_commits` record is deleted.

We only need to delete the records in the
`merge_request_commits_metadata` table when a project is deleted.
We do that in `Projects::DestroyService` just like what we do
for `merge_request_diff_commits` and `merge_request_diff_files`.

In a single project, there can be a lot of records in the
`merge_request_commits_metadata` tied to it. We don't want to
rely on `ON DELETE CASCADE` as that can timeout when deleting
lots of rows (we experienced this with `merge_request_diff_commits`
and `merge_request_diff_files`). So we are following the same
approach of deleting records in batches of 1K.

This is behind `merge_request_diff_commits_dedup` feature flag.