Update project runner feature specs
This change updates the runner feature specs that change in preparation
for the rollout of the vue_project_runners_settings feature flag.

Adds tests for the new UI where necessary and moves old tests to
contexts where the feature flag is disabled.

It also fixes some minor bugs discovered when running the tests.