Change the logic of setting `created_at` for security scans
Previously, scans inherited the `created_at` timestamp from the
associated pipeline, which could be far earlier than we start creating
the security scan.

This could cause the `security_findings` partitions to be dropped
prematurely, as we are using that value to decide if we need to detach
the partition or not.

Now, we use the current time for the first scan in a pipeline and reuse
that timestamp for subsequent scans to avoid this issue.

Changelog: fixed
EE: true