Fix subscription_breakdown_spec.js under Vue 3
This spec was mutating the `state` object created by `createState` by
calling the mutation and passing that. That works under Vue 2, since the
object is made reactive, so the reference is the same.

Under Vue 3, however, a proxy is created internally in Vuex, and the
spec doesn't have access to that. So the mutation has no effect.

Instead, this relies on the `removeLicense` action being triggered by
the modal's primary event.

This also revealed that the implementation doesn't work as the tests
expected, since an error instance is passed to `createAlert`, and not a
string. This also means that two tests cannot pass in reality, since an
error instance is always truthy, and it'll always be different error
instances each time. So, those tests were removed.

We could fix the implementation here, but it's out of scope for fixing
Vue 3 specs.

This also fixes the `createState` calls to pass `licenseRemovalPath`
instead of `licenseRemovePath`, so that the store state is set up
correctly.

One more thing this fixes is the spreading of the `initialStore` value
into the `new Vuex.Store` call. That had no effect, other than
implicitly calling `Vue.use(Vuex)`, because `initialStore` is a function
rather than an object.