Use Gitaly to handle reference existance checks
Contributes to https://gitlab.com/gitlab-org/gitlab/-/issues/556674

**Problem**

Currently GitLab relies on the Redis cache to identify the presence of
branches and tags in the repository. Maintaining this cache is expensive
for large projects.

**Solution**

Trigger a `ListRefs` call to Gitaly to load all necessary branches/tags
per project. Use results of this call for branch/tag presence checks.