Remove unused sorting option
Due to `sort_value_stars_desc` being defined later in the same hash,
this option was over-ridden and thus never used. The code has been like
this for 3 years, and no one noticed or complained, and adding it back
now is essentially a new ~feature::addition, so instead let's address
the bug by deleting it from the codebase.