Describe unit tests for static schema testing
We previously said that unit tests were deprecated, however, we want to
use unit tests for a limited purpose - to statically verify certain
aspects of the schema. For resolving queries and mutations we want to
use integration tests.

This also updates some of our advice around integration testing to
remove the mention of some lesser used methods.