Make gdk-update failures fail pipelines
This makes `gdk-update` CI job failures cause pipelines to fail.

Since the addition of this job, it had a success rate of 99.12% among
about 6,000 runs. If we ignore expected failures and GitLab
infrastructure failures, which also impact other jobs, we even get a net
stability of 99.67%.

Due to the high success and thus low flakiness rate, the `gdk-update`
job can safely allowed to fail, so it can prevent defects from
impacting the development flow of contributors.