Respect `FILTER_ITEM_LIMIT` in autolink filter
When generating links in commit messages, large commits with many links
can cause timeouts. This change ensures no more than
`FILTER_ITEM_LIMIT` links are processed. This change also introduces a
minor refactoring of `StringRegexMarker` so it is possible to check
the number of links to be marked before they are actually marked for
replacement.