Add include_descendant_work_items arg
This MR introduces the include_descendant_work_items argument on the 
work item query, which aims to match the behaviour of the 
include_subepics argument on the existing issues queries.

The arg for work items follows a similar approach to the existing 
include_subepics filter here, which descends the hierarchy recursively 
to fetch the parents before filtering on issues.

Note that there's a slight change in behaviour for work items, where we 
descend the hierarchy regardless of the work item types. For the legacy 
issues endpoint, we previously only descended through the epics.

Changelog: added