Fix generation of change title system note to not be out of range
When issuable title changes and only some parts of it actually
changed (e.g. `Test "issue (1)"` to `issue (1)`), it'll fail to
generate diff and result to a 500 error when updating a title of
MR/issue. Whenever a change like this happens, a `RangeError` gets
raised.

This is because we are passing the HTML escaped title as the raw
title to the `Gitlab::InlineDiffMarker`. That class inherits from
`Gitlab::StringRangeMarker`, expects a `raw_line` and a `rich_line`.
In `Gitlab::InlineDiffMarker`, if we don't pass `rich_line`, we
pass the `raw_line` instead. This is affecting the calculation done
in `Gitlab::StringRangeMarker#mark` which results to the range error.

To fix that, we pass the unescaped title as `line` and it'll be
passed as `rich_line` in `Gitlab::InlineDiffMarker#initialize`.
In `Gitlab::StringRangeMarker#initialize`, it'll be then HTML
escaped. This ensures that the calculation and mapping of range
from raw to rich is working.

Changelog: fixed