Remove unused CDot billing account request
This client endpoint was added to support redirecting only existing
customers to CustomersDot for purchases. We now support purchasing for
all customers, and redirect in all cases so this check has been removed.

Remove the endpoint that the check relied on, as it now has no callers.

Issue: https://gitlab.com/gitlab-org/gitlab/-/issues/472721