Use named chunks in Webpack to ensure deterministic assets generation
In https://gitlab.com/gitlab-com/gl-infra/production/-/issues/20271 we
found that for two different auto-deploy versions, a Webpack chunk file
had identical filenames but they only differed by a single number at the
end of the file.

One file:

```
}},[[347,1,0,4,69,90,126]]);
```

The other had 127:

```
}},[[347,1,0,4,69,90,127]]);
```

That final nested array is part of Webpack’s `webpackJsonp.push` call in
Webpack 4’s output format. It lists:

```
[[<moduleId>, <parentChunkId>, ... ]]
```

In this example, it appears the chunk ID changed from 126 to 127.

In Webpack 4, the sorting criteria is not deterministic if multiple
chunks have the same “weight” (priority). It often falls back to
insertion order, which depends on:

- The order the filesystem returned files for scanning.
- The order async imports were discovered while parsing modules.
- The traversal order through Set (which depends on how chunks were
  added).

To fix this problem for Webpack we need to use `named` chunks. However,
we need a custom function as well to ensure that unnamed async chunks
are sorted in a deterministic way.

Changelog: changed