Pushes to the correct stable branch on AIGW
The existing logic creates tags based on AIGW based on main, not on a
specific version. Which means that when we created tag self-hosted
v17.1.1, it's not using a stable branch, it's just picking up the latest
commit. The updated script will instead create a stable branch if it
doesn't exist yet.