Re-add @vue/compiler-sfc ESM patch
In https://gitlab.com/gitlab-org/gitlab/-/merge_requests/198833, the
patch against the `compiler-sfc.esm-browser.js` file was unintentionally
dropped. This re-adds the patch.

It's not obvious yet whether we need this file to be patched in addition
to the `compiler-sfc.cjs.js` build, but we're adding it back for
consistency.

In this commit, there is only a single Vue SFC using
`<template functional>`, which is
`app/assets/javascripts/diffs/components/diff_row.vue`, so either way,
the impact is small.

This also makes the code comments the same across the two files,
referencing the original merge request
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/114730.