Add event to track AWS Secrets Manager usage
In this MR -
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/191780, we added
logic to track the usage of AWS Secrets Manager in CI/CD pipelines.
But it was not done correctly as it did not define the event and
the metric definition was only tracking monthly usage. We need to track
both weekly and monthly usage

This change fixes the issue by adding the missing
event definition and adding the weekly metric definition.

[Issue](https://gitlab.com/gitlab-org/gitlab/-/issues/542348)

Changelog: changed
EE: true