Add notification email for verification code
We have added a verification code email that will be sent to the user.
In some cases, the user can request to have this email sent to their
secondary address. When this happens, we want to send an email to their
primary email to let them know an email was sent to their secondary
address.