Return old_path when new_path is nil
As part of a DB space saving approach, we are exploring NOT saving new_path
when old_path == new_path. This change supports that, but assuming that
when new_path.nil?, the path has not changed from old_path, and thus
old_path is the correct value to send. This allows us to more seamlessly
drop in a nullification of new_path, rather than having to change our
code-wide expectations of what a nil new_path value means.