Sync validate text limit p_ci_job_artifacts.exposed_as
Previously we introduced the new column 'exposed_as'
to p_ci_job_artifacts and put a text limit constraint on it
in NOT VALID state. It was prepared to be asynchronously
validated on GitLab.com over the weekend.

It has now been confirmed validated on all partitions.
This MR sync validates the contraint.