Skip gdk-update job on backport MRs
This skips the `gdk-update` CI job on merge requests targeting a stable
branch.

We don't need this job on stable branches because it will always use the
latest database from `master` to validate that updates still work. And
updating a GDK installation to a stable branch is currently not in scope
for GDK.