Remove snapshots from security dashboard integration tests
This refactors the security dashboard integration tests in the following
way:

1. Go from snapshot based approach, to checking if the correct
component is loaded
2. Change the structure of the spec so it follows the group, project,
and instance levels
3. Check for the non-configured empty state as well

It also fixes a bug for the not configured project level component. The
`newVulnerabilityPath` and `canAdminVulnerability` properties were not
being provided. This caused an error because it is expected. And this
also caused the Submit vulnerability button not never show up.

Changelog: fixed
EE: true