Init topbar Vue app
For now, this shares the `sidebarData` prop with the SuperSidebar.
This allows us to move over child components without breaking the
current behavior (`global_topbar` disabled).

As soon as we remove the feature flag, we should split and refactor this
prop!

Only the brand logo and user dropdown are moved in this commit. Other
components will move in later commits.