Validate only kubernetes objects using JSON schema
Unlike previously, when the sequence in the array was also validated.
This is done because JSON schema cannot handle validate an array for
relative ordering of elements when there are arbitrary amount of items
present. In other words, unless we do not know the exact amount of items
order of the items in an array cannot be validated.