Add support to Workload for requesting CI Variables explicitly
The `Ci::Workloads::Workload` model (and the service that constructs it)
is a thin wrapper around `CreatePipelineService`. All it does is take a
narrow set of things that could be used in a CI Job and create a
pipeline with just that one jobs details.

When we created `Ci::Workloads::Workload` for Duo Workflow one of the
key requirements was to minimize the risk of extra credentials being
available to the AI. In order to do this we basically passed through
none of the normal CI Variables into the pipeline.

Now as part of
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/200193 we're
trying to make use of the same workload model but pass through some
secret variables (e.g. `ANTHROPIC_API_KEY`). We explored different ways
to pass the variables in but it seems the most appropriate would be to
make use of our existing CI Variables settings which allow for security
feaures like `masked` and `protected`.

This MR introduces a new way to mark specific variables as related to
the `Workload` model when we create a workload. Once the build is picked
up by a runner we then include the narrow subset of variables that were
requested by the workload while still adhering to all our rules around
how variables get passed to jobs.

Most of the logic for this belongs in the
`lib/gitlab/ci/variables/builder.rb` code as this is where we apply
different rules about which variables get passed into workloads. So you
may wish to start there when reviewing it.