Introduce container queries
This is the very first iteration for introducing container queries based
layouts within the product. The long-term purpose behind this change is
for us to be able to move some pages' contents within side-by-side
panels in a single page, which implies that the responsiveness should be
implemented relative to the container panel rather than the whole
viewport.

At first, we'll be migrating individual pages to container queries by
using some sort of "fallback container": the existing `main`
element. This lets us reproduce a behavior that's similar to existing
media queries layouts, with the added benefit that we take the actual
available space into account (eg we better react to sidebars being
expanded).

The idea here is to migrate all existing page to container queries
without fundamentally changing the layout. Once our pages all support
container queries, we should be able to switch to the panels layout
without too much hassle.

Changelog: changed