Display unauthorized projects in admin area projects list page
When the current user has a custom admin role that grants
read_admin_projects permission they should be able to see _all_ projects
in the admin area projects list page. Unauthorized projects (the
user is not a member of) are returned as ProjectMinimalAccess type from
the projects GraphQL endpoint that includes only a very minimal set of
project metadata.