Scope deployment lookup to project in DeploymentsController#show
A deployment belongs to both a project and an environment so either
association could be used to fetch this record. However, because the
IID is defined at the project level, looking up via project is a more
efficient query as it can use the unique index on (project_id, iid).

Changelog: performance