Trigger managed resource deletion when a stop job completes
When an environment has a stop job defined, it goes into a `stopping`
state while the job executes. This meant that managed resource deletion
was skipped, as it only executed for environments that stop immediately
(that is, those with no stop job defined). To support deletions for
environments that do have stop jobs, we trigger deletion explicitly when
the stop job suceeds.

Changelog: fixed