Switch to gitlab_main_user
As we already have a non-null user_id
this is a quick way to shard this table.

It is more preferrable to use gitlab_main_cell,
but the polymorphic `source_id` and `source_type`
means that it will take a while to perform the backfill
via source.
So we shard via user_id instead.