Validate user is owner of provisioned group of trigger service account
Flow triggers are a new AI feature we are adding which is described in
https://gitlab.com/gitlab-org/gitlab/-/issues/559051 .

Flow triggers work by choosing a service account that will then run a
pipeline when a certain action is performed. See
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/200571#note_2673998362
for an early demo.

In an effort to ensure that users cannot gain extra privileges by
assigning service accounts with higher permissions to a flow trigger we
want to limit the user to only select a service account if they are able
to manage that service account. That is equivalent to checking that they
are able to manage service accounts in the group that the service
account was created in.

For this we use `provisioned_by_group` on the service account as and
check if the current user can admin service accounts in that group.

This also relates to
[this Slack thread (internal)](https://gitlab.slack.com/archives/C098GU1HYGG/p1754989399998159?thread_ts=1754902782.469459&cid=C098GU1HYGG).