Make developer role explicitly required for AI flow trigger
Previously we relied on `can?(:create_pipeline)` to authorize access to
run triggers. That's because they create pipelines. But we may still
want to be a little more strict in this case as they are a special case
of pipeline that are getting a service account for `CI_JOB_TOKEN`. In
this special case we want to be sure the user is at least a developer in
this project. That is partially covered by `create_pipeline` but then
there are special edge cases related to creating pipelines that aren't
relevant to this feature.