Update build entity to not use job processing data, artifact_paths
The `artifacts_paths` value is read from the job's metadata record.
This is processing data that may not persist in the future, either
from pipeline archival or otherwise.

As such, we replace the check on `artifacts_paths` with a check on
whether or not the job actually has an `archive` artifact to download.
This is an equivalent replacement.