Set limits for Projects::CompareController#show
When comparing branches with a large number of commits,
the controller was unconditionally fetching pipeline
information for all commits, even though this information
isn't displayed in the UI when there are too many commits.
This was causing timeouts for comparisons with hundreds
of thousands of commits.

This change modifies the controller to only fetch pipeline
information when the number of commits is below
MergeRequestDiff::COMMITS_SAFE_SIZE,
which matches the threshold used for displaying commits in the UI.

The issue: https://gitlab.com/gitlab-org/gitlab/-/issues/550847+