Early deduplication of dependency graph paths
Fixes: https://gitlab.com/gitlab-org/gitlab/-/issues/560154

Prior to this change, we collect all seen graph paths in memory and
deduplicate at the very end of the process. In extreme cases, this can
mean we hold very large numbers of graph paths in memory before
throwing most of them away through deduplication.

This change does deduplication in exactly the same way as before, but
does it when we collect rather than at the end of processing.

I didn't use a Set here as I encountered N+1 issues when adding eql?
and hash methods on the GraphPath object. We can follow up with that
change afterwords.

Changelog: fixed
EE: true

fix