Enable committer_email feature flag to true
This means when the rails app asks for commit data it will also look for
the committer_email. This was behind a feature flag because it was a new
behavior in Gitaly. It has been enabled in production for over a day and
now this commit changes default_enabled to true.

Changelog: added