Do not show heading anchor links in Duo chat messages
We use GLFM to render markdown for Duo chat messages, which by default
provides anchor links to headings.

However, it is not possible to link another user to a Duo chat message,
and it isn't even feasible to share/bookmark a link for yourself to one.
These links should not be shown in Duo chat messages.

They are, anyway, positioned in such a way that they cannot be
interacted with currently. They are a total no-op.

Changelog: fixed