Restrict the issues settings page to root groups
Previously, the issues settings page was
accessible for both root groups and subgroups,
creating inconsistency, as this functionality
should only be available at the root group level.

This commit adds a check to ensure that the page
(including settings for custom fields and statuses)
is only accessible when the group is a root group.

Changelog: fixed
EE: true