Update product_group for Vault usage metrics
Prior to this change, we had Vault usage
metric definitions where the `product_group` was
incorrectly set to `environments` whereas the
correct value should be `pipeline_security`.

This change fixes the issue by updating the
`product_group` for the affected metrics to
the correct value of `pipeline_security`.

Issue
https://gitlab.com/gitlab-org/gitlab/-/issues/521772

Changelog: changed
EE: true