Remove superfluous calls to access_token
These calls were not used to send a Cloud Connector
token in a request, but to check that the caller has
"access to Cloud Connector" in some form.

This doesn't really work, because:

- Cloud Connector is not a feature. Permission checks
  must happen on a per-feature basis using policy checks.
- On gitlab.com, tokens are created on-the-fly, so these
  checks were fruitless because a token always exists,
  and only added extra latency.

On SM/Dedicated instances, where we sync in tokens,
there may now be cases where an upstream call is made
with a missing token. This would still result in a 401
with the same outcome for the customer, however.

Changelog: changed
EE: true