Strip leading whitespace from coverage.xml
This adds a command to the `rspec:coverage` CI job to strip the leading
whitespace from the `coverage.xml` file.

Whitespace in that file constitutes around 6 MB. Stripping it should be
safe because the XML standard doesn't require indentation.

This reduces the size of the `coverage.xml` to around 16 MB, which is
only 6 MB over the default limit instead of 12 MB over the limit.

We could save another ~0.5 MB by removing newlines but we can do that in
a follow-up.