Removes an invalid resolver_complexity method
Whilst investigating how the GraphQL resolvers worked, I stumbled
upon this bug where the complexity isn't actually altered when a
valid query is processed.

This is due to the differences in casing between labelName
which is referenced in both the code and the test and the
actual casing `label_name` which is present when a query
is actually performed against this API.

I've opted to remove this code and the corresponding
invalid unit test so as to avoid making what could be
considered a breaking API change to our GraphQL setup.

This is a nice win as well because less code is less
code.