Fix rspec:coverage not getting any data from fail-fast jobs
Previously it was possible for the `rspec:coverage` not to have any code
coverage artifacts even if `rspec fail-fast` and `rspec-ee fail-fast`
ran because earlier CI rules for `rspec:artifact-collector ee remainder`
matched with `when: never`.

For example, if a merge request were not approved,
`.rails:rules:ee-only-migration` would prevent the artifact collector
job from running.

Fix this by moving the `.rails:rules:rspec fail-fast` rule first so that
it will be created if the fail-fast jobs run.