Fix release environment script spec
In the spec for `release_environments_model.rb`, in the case of branch
pipeline, `CI_COMMIT_REF_NAME` also needs to be stubbed, since it is
also available. Without doing it, when the same spec runs on a tag
pipeline, `CI_COMMIT_REF_NAME` is set to the tag and the spec is broken.

Ref: https://gitlab.com/gitlab-com/gl-infra/delivery/-/issues/21460