Update `anti-abuse` CODEOWNER rules
Anti-Abuse is an old team at GitLab. It is now part of Authorization. 
However, the code review rules don't reflect this completely. 
There was an old group 
`gitlab-org/modelops/anti-abuse/identity-verification-approvers` that
is still a CODEOWNER. That group does not contain all the members of
Authorization; four members vs. nine.

This commit retains the `Identity Verification` section, but shifts
ownership to Authorization.

Resolves https://gitlab.com/gitlab-org/gitlab/-/issues/562261

References:

- https://gitlab.com/groups/gitlab-org/software-supply-chain-security/authorization/approvers/-/group_members?with_inherited_permissions=exclude
- https://gitlab.com/groups/gitlab-org/modelops/anti-abuse/identity-verification-approvers/-/group_members?with_inherited_permissions=exclude