gitaly: Don't parse Git version
A subsequent commit will update GITALY_SERVER_VERSION to a revision of
Gitaly [1] which consumes the `master` branch of Git. When we build Git
from its master branch, it no longer makes sense to report the Git
version as a semver version string such as `2.50.1`. It's more
appropriate to report the commit SHA on its own or something like
`master-<SHA>`.

Remove the Gitlab::Git::Version module as its sole responsibility was
to return a parsed version string. Change the caller so that it asks for
the raw version instead.

[1] https://gitlab.com/gitlab-org/gitaly/-/merge_requests/8030
[2] https://gitlab.com/gitlab-org/gitlab/-/merge_requests/201684