Update user.organizations.first to use new user model association
Previously, the code was using user.organizations.first to get the
first organization from a user's list of organizations. Now it's
been changed to use user.organization directly, which is a more
straightforward way to get a user's primary organization. Instead
of asking for "the first item from the user's organizations list,"
it now simply asks for "the user's organization."