Fix issue promotion when reaching max hierarchy depth
This fixes a bug when use_data_sync_for_issue_promotion is enabled and
promoting an issue to an epic, where the hierarchy limit would be
reached.

It uses a similar pattern as we already have it for the WorkItem
Callbacks, and returns an error on the service.

Additionally, it moves the DataSync::Hierarchy work into the
after_create callback, so it gets executed within a transaction and
rolls back the work item creation, when the hierarchy would not be
valid.